#5134: add test
