+++ src/components/quickBar/QuickBarApp.tsx
@@ -18,17 +18,15 @@
 import React, { useEffect } from "react";
 import ReactDOM from "react-dom";
 import {
-  type Action,
   KBarAnimator,
+  KBarPortal,
   KBarPositioner,
   KBarProvider,
-  KBarPortal,
   KBarSearch,
   useKBar,
   VisualState,
 } from "kbar";
 import ReactShadowRoot from "react-shadow-root";
-import quickBarRegistry from "@/components/quickBar/quickBarRegistry";
 import faStyleSheet from "@fortawesome/fontawesome-svg-core/styles.css?loadAsUrl";
 import { expectContext } from "@/utils/expectContext";
 import { once } from "lodash";
@@ -39,6 +37,7 @@ import selection from "@/utils/selectionController";
 import { animatorStyle, searchStyle } from "./quickBarTheme";
 import QuickBarResults from "./QuickBarResults";
 import useActionGenerators from "@/components/quickBar/useActionGenerators";
+import useActions from "@/components/quickBar/useActions";
 
 /**
  * Set to true if the KBar should be displayed on initial mount (i.e., because it was triggered by the
@@ -50,35 +49,6 @@ let autoShow = false;
  * Window event name to programmatically trigger quick bar
  */
 const QUICKBAR_EVENT_NAME = "pixiebrix-quickbar";
-
-function useActions(): void {
-  // The useActions hook is included in KBarComponent, which mounts/unmounts when the kbar is toggled
-
-  const { query } = useKBar();
-  const uninstallActionsRef = React.useRef<() => void | null>(null);
-
-  // Listen for changes while the kbar is mounted:
-  // - The user is making edits in the Page Editor
-  // - Generators are producing new actions in response to the search query changing
-  useEffect(() => {
-    const handler = (nextActions: Action[]) => {
-      uninstallActionsRef.current?.();
-      // Potential improvement: to avoid flickering, we could register actions individually and keep track of
-      // their uninstall handlers by id.
-      uninstallActionsRef.current = query.registerActions(nextActions);
-    };
-
-    // Don't use useRegisterActions, because then we aren't able to unregister actions that were around
-    // from the initial mount
-    handler(quickBarRegistry.currentActions);
-
-    quickBarRegistry.addListener(handler);
-    return () => {
-      quickBarRegistry.removeListener(handler);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps -- the query is available on initial mount
-  }, []);
-}
 
 function useAutoShow(): void {
   const { query } = useKBar();
